Chore: 대학 목록 리스트 조회에서 국가랑 교환유형 타입 변경

+++ src/main/java/com/likelion/boomarble/domain/universityInfo/dto/UniversityInfoDetailDTO.java
@@ -16,10 +16,10 @@ public class UniversityInfoDetailDTO {
     private String country;
     private String period; // 파견기간
     private int recruitNum; // 선발인원
-    private int gradeQ; // 지원자격 학점
+    private float gradeQ; // 지원자격 학점
     private int ibtQ; // 지원자격 ibt
     private int toeflQ; // 지원자격 토플
-    private int ielts; // 지원자격 ielts
+    private float ielts; // 지원자격 ielts
     private String japaneseQ; // 지원자격 jlpt
     private String chineseQ; // 지원자격 hsk
     private String qualificationEtc; // 지원자격 기타
